use big integer and decimal values for all literals. This breaks stuff, so some tests and documentation needs to be updated.

+++ src/main/java/com/inksetter/twist/expression/DoubleLiteral.java
@@ -3,17 +3,19 @@ package com.inksetter.twist.expression;
 import com.inksetter.twist.EvalContext;
 import com.inksetter.twist.Expression;
 
+import java.math.BigDecimal;
+
 public class DoubleLiteral implements Expression {
-    public DoubleLiteral(Double value) {
+    public DoubleLiteral(BigDecimal value) {
         _value = value;
     }
 
-    public Double evaluate(EvalContext ctx) {
+    public BigDecimal evaluate(EvalContext ctx) {
         return _value;
     }
 
     @Override
     public String toString() { return _value.toString(); }
 
-    private final Double _value;
+    private final BigDecimal _value;
 }
